EDIT detail
